feat(core): rename of API SCD core function with the version number
+ add open scd event
