+++ packages/core/src/scd-events-v1/index.ts
@@ -1,1 +1,2 @@
-export * from './scd-edit-events'
+export * from './edit.scd-events-v1'
+export * from './open.scd-events-v1'
