--- packages/core/src/scd-events/index.ts
@@ -1,1 +1,0 @@
-export * from "./scd-edit-events"
